requirements.txt
fastapi==0.115.3
mariadb==1.1.10
paho_mqtt==2.1.0
pip==24.2
pydantic==2.9.2
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.30.6
